requirements.txt
flask>=2.0.0,<3.0.0
Werkzeug<3.0.0
python-dotenv>=0.19.0
requests>=2.25.0
jinja2>=3.0.0
